+++ src/main/kotlin/fr/xpdustry/javelin/util/utils.kt
@@ -1,8 +1,11 @@
+@file:JvmName("Utils")
+
 package fr.xpdustry.javelin.util
 
 import cloud.commandframework.services.ServicePipeline
 import com.google.gson.Gson
 import fr.xpdustry.distributor.Distributor
+import fr.xpdustry.distributor.command.sender.ArcCommandSender
 import fr.xpdustry.distributor.message.MessageIntent
 import fr.xpdustry.distributor.message.format.MessageFormatter
 import io.leangen.geantyref.TypeToken
@@ -12,18 +15,11 @@ import org.java_websocket.protocols.Protocol
 
 private val EMPTY_ARRAY = emptyArray<Any>()
 
-var clientMessageFormatter: MessageFormatter
-    get() = Distributor.getClientMessageFormatter()
-    set(value) = Distributor.setClientMessageFormatter(value)
-
-/*
-var serverMessageFormatter: MessageFormatter
-    get() = Distributor.getServerMessageFormatter()
-    set(value) = Distributor.setServerMessageFormatter(value)
- */
-
 val servicePipeline: ServicePipeline
     get() = Distributor.getServicePipeline()
+
+val ArcCommandSender.formatter: MessageFormatter
+    get() = Distributor.getMessageFormatter(this)
 
 fun MessageFormatter.format(message: String, intent: MessageIntent = MessageIntent.INFO): String =
     format(intent, message, EMPTY_ARRAY)
@@ -31,15 +27,6 @@ fun MessageFormatter.format(message: String, intent: MessageIntent = MessageInte
 inline fun <reified T> typeToken(): TypeToken<T> = object : TypeToken<T>() {}
 
 inline fun <reified T> Gson.fromJson(json: String): T = fromJson(json, typeToken<T>().type)
-
-/*
-fun getHttpsDraft(): Draft {
-    val protocols = mutableListOf<IProtocol>()
-    protocols += Protocol("ocpp2.0")
-    protocols += Protocol("")
-    return Draft_6455(emptyList(), protocols)
-}
- */
 
 fun getDraft(https: Boolean): Draft {
     return if (https) {